### 2025-01-28 #7731 @emgeopaula
Hi @kthyng, thank you for handling my submission.

Here are my reviewer suggestions and their email/github contacts:
Xiao Longying	longying.xiao@gtk.fi
Pankaj K Mishra 	pankajkmishra
Christopher Edsall 	christopheredsall
Joe Wallwork 	jwallwork23
Andy Nowacki 	anowacki
Haipeng Li 	Haipeng-ustc

I see that the editorialbot has detected that my paper is 54 words too long and that some DOIs are missing. Shall I make the changes directly in the submitted paper.md file?
