### 2025-01-28 #7731 @kthyng
@emgeopaula Thank you for paying attention to the word limit! 54 words over is close enough. Please update the DOIs in your branch JOSS_paper version of paper.md.
