+++ paper.bib
@@ -5,6 +5,7 @@
   VOLUME = {119},
   YEAR = {2018},
   PAGES = {126--136},
+  doi = {https://doi.org/10.1016/j.cageo.2018.07.005}
 }
 
 @article{Castillo-Reyes2023,
@@ -35,7 +36,8 @@
   pages = 	 {486--514},
   OPTmonth = 	 {November},
   OPTnote = 	 {},
-  OPTannote = 	 {}
+  OPTannote = 	 {},
+  doi = {10.1111/j.1365-246X.2008.03902.x}
 }
 
 @Article{kals10,
@@ -49,7 +51,8 @@
   pages = 	 {1174--1188},
   OPTmonth = 	 {September},
   OPTnote = 	 {},
-  OPTannote = 	 {}
+  OPTannote = 	 {},
+  doi = {10.1111/j.1365-246X.2010.04686.x}
 }
 
 @Article{kals12,
@@ -63,7 +66,8 @@
   pages = 	 {E301--E315},
   OPTmonth = 	 {July--August},
   OPTnote = 	 {},
-  OPTannote = 	 {}
+  OPTannote = 	 {},
+  doi = {10.1190/geo2012-0018.1},
 }
 
 @Article{kals15,
@@ -77,7 +81,8 @@
   pages = 	 {1429--1452},
   OPTmonth = 	 {September},
   OPTnote = 	 {},
-  OPTannote = 	 {AMT, CSAMT, TEM, resistivity, seismic velocity, petrophysics}
+  OPTannote = 	 {AMT, CSAMT, TEM, resistivity, seismic velocity, petrophysics},
+  doi = {10.1093/gji/ggv184}
 }
 
 
@@ -89,7 +94,8 @@
   number={2},
   pages={700--718},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/gji/ggt15}
 }
 
 
@@ -100,14 +106,15 @@
   VOLUME = {227},
   YEAR = {2021},
   PAGES = {1624--1645},
-  NOTE = {\href{https://doi.org/10.1093/gji/ggab264}{doi: 10.1093/gji/ggab264}}
+  doi = {10.1093/gji/ggab264}
 }
 
 @phdthesis{Rulff2023,
   author  = "Paula Rulff",
   title   = "Three-dimensional forward modelling and inversion of controlled-source electromagnetic data using the edge-based finite-element method",
   school  = "Uppsala University",
-  year    = "2023"
+  year    = "2023",
+  url =  {https://urn.kb.se/resolve?urn=urn:nbn:se:uu:diva-514339}
 }
 
 @article{Rulff2024,
@@ -129,7 +136,8 @@
   number={1},
   pages={15--41},
   year={2001},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi = {10.1137/s0895479899358194},
 }
 
 @article{Rochlitz2019,
