+++ paper.bib
@@ -95,7 +95,7 @@
   pages={700--718},
   year={2013},
   publisher={Oxford University Press},
-  doi = {10.1093/gji/ggt15}
+  doi = {https://doi.org/10.1093/gji/ggt154}
 }
 
 
