### 2025-02-07 #7731 @emgeopaula
@kthyng I updated the invalid DOI.
I do not see a problem with https://doi.org/10.1016/j.cageo.2018.07.005 as I used the url prefixes for all DOIs.
