+++ paper.bib
@@ -5,7 +5,7 @@
   VOLUME = {119},
   YEAR = {2018},
   PAGES = {126--136},
-  doi = {https://doi.org/10.1016/j.cageo.2018.07.005}
+  doi = {10.1016/j.cageo.2018.07.005}
 }
 
 @article{Castillo-Reyes2023,
@@ -95,7 +95,7 @@
   pages={700--718},
   year={2013},
   publisher={Oxford University Press},
-  doi = {https://doi.org/10.1093/gji/ggt154}
+  doi = {doi.org/10.1093/gji/ggt154}
 }
 
 
