### 2025-03-26 #7949 @editorialbot
[REVIEW]: elfe3D: Modelling with the total electric field approach using finite elements in 3D
**Submitting author:** @emgeopaula (<a href="http://orcid.org/0000-0001-6714-8008">Paula Rulff</a>)
**Repository:** https://github.com/emsig/elfe3D
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v1.0.1
