### 2025-04-01 #7949 @pankajkmishra
@AnjaliSandip  I have finished the review. My checklist is complete! The documentation in the PDF manual is very good but I could not check detailed code APIs using doxygen

- Installation was easy on and as suggested on

```
LSB Version:    core-11.1.0ubuntu2-noarch:printing-11.1.0ubuntu2-noarch:security-11.1.0ubuntu2-noarch
Distributor ID: Ubuntu
Description:    Ubuntu 20.04.6 LTS
Release:        20.04
Codename:       focal
```

@emgeopaula It appears that when OpenMP is allowed to use all threads by default, the performance drops drastically (never finishes execuitng the code)—likely due to oversubscription. Perhaps there should be a comment about setting export OMP_NUM_THREADS  to a suitable number. I could only execute the code after manually exporting OMP_NUM_THREADS= (less than total number of CPUs I have). I look forward to see inversion examples in this repo in the near future.
