+++ paper.bib
@@ -95,7 +95,7 @@
   pages={700--718},
   year={2013},
   publisher={Oxford University Press},
-  doi = {doi.org/10.1093/gji/ggt154}
+  doi = {10.1093/gji/ggt154}
 }
 
 
