### 2025-04-08 #7949 @emgeopaula
Hi @pankajkmishra and @jwallwork23,

thank you for agreeing to review my code and software paper!

@pankajkmishra, thank you for your sugggestions. It is very helpful that you are reporting the OpenMP issue. I will add a comment about setting export OMP_NUM_THREADS to a suitable number. I will also upload an HTML version of the Doxygen output to the repository.
