### 2025-04-09 #7949 @jwallwork23
Hi @emgeopaula. Some initial thoughts and requests upon reading the paper.

* The "Statement of Need" section seems to last the whole paper? It would be good if the paper was divided into other sections. In particular, there should be a section that describes the state of the field and how the software compares to other commonly-used packages.
* I'm not sure that the description of Maxwell's equations really fits in the statement of need section. It's also not clear to me that the equation is needed in a JOSS paper? (Although maybe other reviewers and the editor have other opinions?). I note that the equation and notation are introduced but then not referred to again. Indeed, I also note that the Maxwell equations are not included in the similar section of the README.
* Interesting that you mention *forward* modelling and *goal-oriented* mesh adaptation, but not *adjoint* modelling. Surely this is required for goal-oriented error estimation and mesh adaptation? Is a continuous or discrete adjoint approach being deployed? If the latter, what kind of automatic differentiation are you using?
* Could you clarify what the "Ex" field is mentioned in Figure 1?
