### 2025-04-09 #7949 @emgeopaula
> [@emgeopaula](https://github.com/emgeopaula) I could produce html files but did not find contents in them. I might be doing something wrong but perhaps you could mention what one should expect from doxygen generated files.
>
> Adding html directly or a GitHub page style documentation could be another idea. The PDF manual is quite good
>
> [@AnjaliSandip](https://github.com/AnjaliSandip) From my side this submission is good to go
