### 2025-04-09 #7949 @emgeopaula
@pankajkmishra Thanks for your review. The doxygen generated files extract information on code routine descriptions, code structure and dependencies. For convenience, I will try to upload the html files to the doc directory.
