+++ docs/README.md
@@ -3,4 +3,4 @@ Here, you find the elfe3D code documentation as pdf and the Kappa (thesis withou
 To generate a more detailed code documentation from source code comments, including information about routines, functions, and variables, you can run doxygen with the provided Doxyfile.
 For more information see: https://www.doxygen.nl/.
 
-For convenience, you find html files of the Doxygen Documentation in /elfe3d_v1.1.0_Doxygen_html. When opening the index.html file, the documentation will appear in your browser window.
+For convenience, you find html files of the Doxygen Documentation in /elfe3d_v1.1.0_Doxygen_html.zip. When opening the index.html file, the documentation will appear in your browser window.
