+++ docs/elfe3d_v1.1.0_Doxygen_html/readme.md
@@ -1,1 +1,0 @@
-Open any of the .html files to view the documentation in your browser.
