+++ docs/README.md
@@ -1,3 +1,5 @@
 Here, you find the elfe3D code documentation as pdf and the Kappa (thesis without papers) of my PhD thesis.
 
-You can use the provided Doxyfile to produce a detailed code documentation.
+To generate a more detailed code documentation from source code comments, including information about routines, functions, and variables, you can run doxygen with the provided Doxyfile. For more information see: https://www.doxygen.nl/.
+
+For convenience, you find html files of the Doxygen Documentation in elfe3d_v1.1.0_Doxygen_html.zip. When opening the index.html file, the documentation will appear in your browser window.
