+++ paper.md
@@ -41,11 +41,11 @@ Standard strategies in electromagnetic modelling often avoid 3D simulations due 
 
 In `elfe3D`, three state-of-the-art functionalities are combined:
 
-1. **Variable Magnetic Permeability**: In traditional 3D electromagnetic forward modeling algorithms, magnetic permeability and dielectric permittivity are typically not considered variable model parameters. This oversight is often due to the relatively minor contrasts in these properties compared to conductivity contrasts prevalent in most geological environments at frequencies in the Hz- to kHz-range. However, it is well documented that mineralised zones and metallic infrastructure [@Heagy2023] may exhibit pronounced contrasts in magnetic permeability relative to their host rocks. Dielectric permittivities influence only high-frequency data in very resistive environments [@kals08]. Consequently, in `elfe3D`, magnetic permeability is treated as a variable model parameter, in conjunction with electrical resistivity.
+1. **Variable Magnetic Permeability**: In traditional 3D electromagnetic forward modeling algorithms, magnetic permeability and dielectric permittivity are typically not considered variable model parameters. This oversight is often due to the relatively minor contrasts in these properties compared to conductivity contrasts prevalent in most geological environments at frequencies in the Hz- to kHz-range. However, it is well documented that mineralised zones and metallic infrastructure [@Heagy2023] may exhibit pronounced contrasts in magnetic permeability relative to their host rocks. Dielectric permittivities influence only high-frequency data in very resistive environments [@kals08]. Consequently, in `elfe3D`, magnetic permeability is treated as a variable model parameter, in conjunction with electrical resistivity. @Rulff2021 present a numerical test example with magnetic anomalies.
 
-2. **Automatic Mesh Refinement**: The refinement approach implemented in `elfe3D` is inspired by @Ren2013. It aims to achieve an optimal balance between the accuracy of the solutions at specified areas of interest and the overall size of the computational problem. This is accomplished through a goal-oriented mesh refinement strategy, specifically tailored for models characterised by variable electrical conductivity and magnetic permeability.
+2. **Automatic Mesh Refinement**: The refinement approach implemented in `elfe3D` is inspired by @Ren2013. It aims to achieve an optimal balance between the accuracy of the solutions at specified areas of interest and the overall size of the computational problem. This is accomplished through a goal-oriented mesh refinement strategy, specifically tailored for models characterised by variable electrical conductivity and magnetic permeability. The goal-oriented mesh refinement strategy including an adjoint problem formulation and the utilised error estimation approaches are detailed in @Rulff2021.
 
-3. **Perfect Electric Conductors**: `elfe3D` incorporates a methodology for approximating highly conductive infrastructures within the model, such as steel-cased wells that are frequently encountered in exploration contexts. These structures are modeled as perfect electric conductors [@Um2020], utilising an approach similar to boundary conditions. Within the computational domain, this formulation enforces the electric field to zero at mesh edges corresponding to the perfect electric conductors.
+3. **Perfect Electric Conductors**: `elfe3D` incorporates a methodology for approximating highly conductive infrastructures within the model, such as steel-cased wells that are frequently encountered in exploration contexts. These structures are modeled as perfect electric conductors [@Um2020], utilising an approach similar to boundary conditions. Within the computational domain, this formulation enforces the electric field to zero at mesh edges corresponding to the perfect electric conductors. @Castillo-Reyes2023 provide details on the perfect electric conductor implementation and application.
 
 
 The program `elfe3D` is designed for geophysicists who want to:
