### 2025-04-09 #7949 @AnjaliSandip
> [@jwallwork23](https://github.com/jwallwork23) Thanks for your comments! I will address points 3 and 4 in an updated version of the paper. Point 3: the code is indeed solving a discrete adjoint problem, similar to the discrete forward problem, for error estimation. The approach is detailed in the cited article: Rulff et al (2021). Point 4: It is the electric field in x-direction.
>
> [@AnjaliSandip](https://github.com/AnjaliSandip) Would you be so kind as to provide your opinion on points 1 and 2 raised by [@jwallwork23](https://github.com/jwallwork23)? Before I restructure the paper, I would like to know if it is necessary. I was writing it with the idea in mind that the paper should contain a section titled 'Statement of need' that briefly describes the mathematical problem the software is solving, state of the art, software capabilities and target audience. I can, of course, restructure and extend the paper w.r.t the points raised by [@jwallwork23](https://github.com/jwallwork23)!

@emgeopaula I agree with @jwallwork23's points 1 and 2.
