+++ paper.md
@@ -29,17 +29,17 @@ The controlled-source electromagnetic method is a geophysical technique that det
 # Statement of need
 
 The numerical modelling program `elfe3D` calculates forward responses, i.e. electric and magnetic field components, for frequency-domain three-dimensional (3D) controlled-source electromagnetic surveys in geophysics [Figure \ref{fig:example}]. 
-The physical behavior of electromagnetic fields is described by Maxwell's equations, which serve as the foundation for the curl-curl equation in terms of the total electric field $\bf{E}$ that `elfe3D` is solving:
+Compared to standard electromagnetic geophysical simulation software, `elfe3D` excels in flexibility regarding subsurface properties and geometries as well as survey settings, i.e. receivers can be arbitrarily placed in the modelling domain and the electrical properties can be flexibly distributed in the subsurface upon model design. Implemented adaptive mesh refinement approaches can automatically design problem-specific meshes and optimise computational load and solution accuracy.
 
-$$\nabla \times \frac{1}{\mu} \nabla \times \mathbf{E} - i\omega \frac{1}{\rho} \mathbf{E} - \omega^2 \epsilon \mathbf{E}  = \color{black} i\omega \mathbf{J}_{p} \color{black} \quad \text{in} \quad \Omega,$$
-
-where $\Omega$ is the computational domain, $\omega$ is the angular frequency, $\rho$ the electrical resistivity, $\varepsilon$ the dielectric permittivity and $\mu$ the magnetic permeability. A time dependence $e^{-i\omega t}$ is assumed and $\textbf{J}_{p}$ describes the source term.  Dirichlet boundary conditions are imposed on the outer domain boundaries.
-
-Isotropic electric resistivities and magnetic permeabilities are variable, but element-wise constant, model parameters arranged in a tetrahedral mesh generated with `TetGen` [@Si2015]. Line or loop sources are represented along the edges of the mesh elements. The governing equations are discretised with linear vector finite-element shape functions [@Jin2014]. Developed using modern Fortran, `elfe3D` leverages vectorisation and shared-memory parallelism and employs a direct method, the `MUMPS` solver [@Amestoy2001], for solving the system of equations.
+# State of the field
 
 Standard strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant in the curl-curl equation at typical frequencies used in controlled-source electromagnetic surveys. Contemporary controlled-source electromagnetic field investigations employ advanced measurement configurations designed to image complex subsurface structures. This context necessitates the development of flexible and manageable 3D simulation software tailored for the evaluation of controlled-source electromagnetic data. The numerical techniques required to facilitate this advancement are sophisticated, and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [@Grayver2015; @Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019].
 
-In `elfe3D`, three state-of-the-art functionalities are combined:
+how this software compares to other commonly-used
+
+# Special code capabilities
+
+In `elfe3D`, three state-of-the-art functionalities are combined. In summary, these features distinguish `elfe3D` from other available codes:
 
[…]
