+++ paper.md
@@ -28,7 +28,7 @@ The controlled-source electromagnetic method is a geophysical technique that det
 
 # Statement of need
 
-The numerical modelling program `elfe3D` calculates forward responses, i.e. electric and magnetic field components, for frequency-domain three-dimensional (3D) controlled-source electromagnetic surveys in geophysics [Figure \ref{fig:example}]. 
+The purpose of the software is to ...The numerical modelling program `elfe3D` calculates forward responses, i.e. electric and magnetic field components, for frequency-domain three-dimensional (3D) controlled-source electromagnetic surveys in geophysics [Figure \ref{fig:example}]. 
 Compared to standard electromagnetic geophysical simulation software, `elfe3D` excels in flexibility regarding subsurface properties and geometries as well as survey settings, i.e. receivers can be arbitrarily placed in the modelling domain and the electrical properties can be flexibly distributed in the subsurface upon model design. Implemented adaptive mesh refinement approaches can automatically design problem-specific meshes and optimise computational load and solution accuracy.
 
 # State of the field
@@ -57,19 +57,25 @@ The program `elfe3D` is designed for geophysicists who want to:
 - Further develop the code.
 
 # Workflow overview
+
+Isotropic electric resistivities and magnetic permeabilities are variable, but element-wise constant, model parameters arranged in a tetrahedral mesh generated with `TetGen` [@Si2015]. The mesh files must be provided as input to `elfe3D` along with a start file specifying survey parameters such as frequencies, receiver and source positions, etc. Line or loop sources are represented along the edges of the mesh elements.
+
 The physical behavior of electromagnetic fields is described by Maxwell's equations, which serve as the foundation for the curl-curl equation in terms of the total electric field $\bf{E}$ that `elfe3D` is solving:
 
 $$\nabla \times \frac{1}{\mu} \nabla \times \mathbf{E} - i\omega \frac{1}{\rho} \mathbf{E} - \omega^2 \epsilon \mathbf{E}  = \color{black} i\omega \mathbf{J}_{p} \color{black} \quad \text{in} \quad \Omega,$$
 
 where $\Omega$ is the computational domain, $\omega$ is the angular frequency, $\rho$ the electrical resistivity, $\varepsilon$ the dielectric permittivity and $\mu$ the magnetic permeability. A time dependence $e^{-i\omega t}$ is assumed and $\textbf{J}_{p}$ describes the source term.  Dirichlet boundary conditions are imposed on the outer domain boundaries.
 
-Isotropic electric resistivities and magnetic permeabilities are variable, but element-wise constant, model parameters arranged in a tetrahedral mesh generated with `TetGen` [@Si2015]. Line or loop sources are represented along the edges of the mesh elements. The governing equations are discretised with linear vector finite-element shape functions [@Jin2014]. Developed using modern Fortran, `elfe3D` leverages vectorisation and shared-memory parallelism and employs a direct method, the `MUMPS` solver [@Amestoy2001], for solving the system of equations. From the solution $\bf{E}$, the synthetic data, that is the electric and magnetic field components in Cartesian space, are calculated and written to output files.
+This governing equation is discretised with linear vector finite-element shape functions [@Jin2014]. Developed using modern Fortran, `elfe3D` leverages vectorisation and shared-memory parallelism and employs a direct method, the `MUMPS` solver [@Amestoy2001], for solving the system of equations. From the solution $\bf{E}$, the synthetic data, that is the electric and magnetic field components in Cartesian space, are calculated and written to output files.
+
[…]
