+++ paper.md
@@ -28,14 +28,13 @@ The controlled-source electromagnetic method is a geophysical technique that det
 
 # Statement of need
 
-Advanced and customisable forward modeling and inversion approaches are required to obtain resistivity models of the subsurface from electromagnetic geophysical data [@Rochlitz2019]. The purpose of `elfe3D` is to numerically calculate forward responses from subsurface models, i.e. electric and magnetic field components, for frequency-domain three-dimensional (3D) controlled-source electromagnetic surveys in geophysics [Figure \ref{fig:example}]. 
+Contemporary electromagnetic geophysical field investigations employ advanced measurement configurations designed to image complex subsurface structures. This context necessitates the development of flexible and manageable three-dimensional (3D) simulation software tailored for the evaluation of controlled-source electromagnetic data [@Rochlitz2019]. To address this task, `elfe3D` was developed to numerically calculate forward responses, i.e. electric and magnetic field components, for 3D subsurface models studied with frequency-domain controlled-source electromagnetic surveys [Figure \ref{fig:example}]. 
 Compared to standard electromagnetic geophysical simulation software, `elfe3D` excels in flexibility regarding subsurface properties and geometries as well as survey settings, i.e. receivers can be arbitrarily placed in the modelling domain and the electrical properties can be distributed in the subsurface upon model design. Implemented adaptive mesh refinement approaches can automatically design problem-specific meshes and balance computational load and solution accuracy.
 
 # State of the field
 
-Standard strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant in the curl-curl equation at typical frequencies used in controlled-source electromagnetic surveys. Contemporary controlled-source electromagnetic field investigations employ advanced measurement configurations designed to image complex subsurface structures. This context necessitates the development of flexible and manageable 3D simulation software tailored for the evaluation of controlled-source electromagnetic data. The numerical techniques required to facilitate this advancement are sophisticated, and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [@Grayver2015; @Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019].
+Standard strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant at typical frequencies used in controlled-source electromagnetic surveys. The numerical techniques required to facilitate 3D simuations and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [Boerner2010] and some open-source codes became available in recent years [@Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019].
 
-how this software compares to other commonly-used
 
 # Special code capabilities
 
@@ -43,7 +42,7 @@ In `elfe3D`, three state-of-the-art functionalities are combined. In summary, th
 
 1. **Variable Magnetic Permeability**: In traditional 3D electromagnetic forward modeling algorithms, magnetic permeability and dielectric permittivity are typically not considered variable model parameters. This oversight is often due to the relatively minor contrasts in these properties compared to conductivity contrasts prevalent in most geological environments at frequencies in the Hz- to kHz-range. However, it is well documented that mineralised zones and metallic infrastructure [@Heagy2023] may exhibit pronounced contrasts in magnetic permeability relative to their host rocks. Dielectric permittivities influence only high-frequency data in very resistive environments [@kals08]. Consequently, in `elfe3D`, magnetic permeability is treated as a variable model parameter, in conjunction with electrical resistivity. @Rulff2021 present a numerical test example with magnetic anomalies.
 
[…]
