+++ paper.bib
@@ -258,3 +258,16 @@ geothermal applications},
   eventdate    = {2024-09-07/2024-09-13},
   venue        = {Beppu, Japan},
 }
+
+@article{Boerner2010,
+author		= {B{\"o}rner, Ralph-Uwe},
+journal		= {Surveys in Geophysics},
+number		= {2},
+pages		= {225--245},
+title		= {Numerical modelling in geo-electromagnetics: {Advances} and challenges},
+volume		= {31},
+year		= {2010},
+language	= {English},
+publisher	= {Springer},
+doi         =  10.1007/s10712-009-9087-x},
+}
