+++ paper.bib
@@ -269,5 +269,5 @@ volume		= {31},
 year		= {2010},
 language	= {English},
 publisher	= {Springer},
-doi         =  10.1007/s10712-009-9087-x},
+doi         =  {10.1007/s10712-009-9087-x},
 }
