+++ paper.bib
@@ -271,3 +271,38 @@ language	= {English},
 publisher	= {Springer},
 doi         =  {10.1007/s10712-009-9087-x},
 }
+
+@article{Oldenburg2020,
+author = {Douglas W. Oldenburg and Lindsey J. Heagy and Seogi Kang and Rowan Cockett and},
+title = {3D electromagnetic modelling and inversion: a case for open source},
+journal = {Exploration Geophysics},
+volume = {51},
+number = {1},
+pages = {25--37},
+year = {2020},
+publisher = {Taylor \& Francis},
+doi = {10.1080/08123985.2019.1580118},
+}
+
+@article{Werthmuller2021,
+    author = {Werthmüller, Dieter and Rochlitz, Raphael and Castillo-Reyes, Octavio and Heagy, Lindsey},
+    title = {Towards an open-source landscape for 3-D CSEM modelling},
+    journal = {Geophysical Journal International},
+    volume = {227},
+    number = {1},
+    pages = {644-659},
+    year = {2021},
+    month = {06},
+    doi = {10.1093/gji/ggab238},
+}
+
+@article{Weiss2023,
+author = {Weiss, Michael and Neytcheva, Maya and Kalscheuer, Thomas},
+publisher = {Computational Geosciences},
+title = {{Iterative solution methods for 3D controlled-source electromagnetic forward modelling of geophysical exploration scenarios}},
+year = {2023},
+journal={Computational Geosciences},
+volume={27},
+pages={81--102},
+doi = {10.1007/s10596-022-10182-2},
+}
