+++ paper.md
@@ -33,7 +33,7 @@ Compared to standard electromagnetic geophysical simulation software, `elfe3D` e
 
 # State of the field
 
-Common strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant at typical frequencies used in controlled-source electromagnetic surveys. The numerical techniques required to facilitate 3D simuations and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [Boerner2010] and some open-source codes that can simulate controlled-source electromagnetic data became available in recent years [@Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019]. Objectives for designing these and other new electromagnetic forward modelling codes are to (i) enable modelling of complex subsurface settigs [@Rochlitz2019], (ii) enhance the accuracy of forward responses while minimising computational costs [@Grayver2015], (iii) optimise and understand the effects of model discretisation [@Oldenburg2020, Werthmuller2021, @Castillo-Reyes2023] and to develop efficient solvers [@Weiss2023b].
+Common strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant at typical frequencies used in controlled-source electromagnetic surveys. The numerical techniques required to facilitate 3D simuations and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [Boerner2010] and some open-source codes that can simulate controlled-source electromagnetic data became available in recent years [@Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019]. Objectives for designing these and other new electromagnetic forward modelling codes are to (i) enable modelling of complex subsurface settigs [@Rochlitz2019], (ii) enhance the accuracy of forward responses while minimising computational costs [@Grayver2015], (iii) optimise and understand the effects of model discretisation [@Oldenburg2020, Werthmuller2021, @Castillo-Reyes2023] and to develop efficient solvers [@Weiss2023].
 
 
 # Special code capabilities
