+++ paper.md
@@ -29,13 +29,13 @@ The controlled-source electromagnetic method is a geophysical technique that det
 # Statement of need
 
 Contemporary electromagnetic geophysical field investigations employ advanced measurement configurations designed to image complex subsurface structures. This context necessitates the development of flexible and manageable three-dimensional (3D) simulation software tailored for the evaluation of controlled-source electromagnetic data [@Rochlitz2019]. To address this task, `elfe3D` was developed to numerically calculate forward responses, i.e. electric and magnetic field components, for 3D subsurface models studied with frequency-domain controlled-source electromagnetic surveys [Figure \ref{fig:example}]. 
-Compared to standard electromagnetic geophysical simulation software, `elfe3D` excels in flexibility regarding subsurface properties and geometries as well as survey settings, i.e. receivers can be arbitrarily placed in the modelling domain and the electrical properties can be distributed in the subsurface upon model design. Implemented adaptive mesh refinement approaches can automatically design problem-specific meshes and balance computational load and solution accuracy.
+Compared to standard electromagnetic geophysical simulation software, `elfe3D` excels in flexibility regarding subsurface properties and geometries as well as survey settings, i.e. receivers can be arbitrarily placed in the modelling domain. Implemented adaptive mesh refinement approaches can automatically design problem-specific meshes and balance computational load and solution accuracy.
 
 ![Key steps of the forward modelling procedure including the choice of a subsurface model and source-receiver setup (Step I; note that the subsurface anomaly and survey setup are enlarged in the image for better visibility), the meshing of the modelling domain (Step II; note that a slice through the inner modelling domain is displayed) and the calculation of field responses with `elfe3D` (Step III; note that only the electric field component in x-direction (Ex) is displayed). Figure adapted from PhD thesis: @Rulff2023.\label{fig:example}](modelling-procedure-elfe3D.png){ width=90% }
 
 # State of the field
 
-Common strategies in electromagnetic modelling often avoid 3D simulations due to their substantial computational demands and the predominant focus tends to be on electrical resistivity models, as this material property is dominant at typical frequencies used in controlled-source electromagnetic surveys. The numerical techniques required to facilitate 3D simuations and their integration into controlled-source electromagnetic modelling code is currently a prominent area of academic research [@Boerner2010] and some open-source codes that can simulate controlled-source electromagnetic data became available in recent years [@Heagy2017; @Castillo-Reyes2018; @Rochlitz2019; @Werthmueller2019]. Objectives for designing these and other new electromagnetic forward modelling codes are to (i) enable modelling of complex subsurface settings [@Rochlitz2019], (ii) enhance the accuracy of forward responses while minimising computational costs [@Grayver2015], (iii) optimise and understand the effects of model discretisation [@Oldenburg2020, Werthmuller2021, @Castillo-Reyes2023] and (iv) to develop efficient solvers [@Weiss2023].
[…]
