### 2025-04-15 emsig/elfe3D#3 @joewallwork
[JOSS Review] Community guidelines
Towards https://github.com/openjournals/joss-reviews/issues/7949.

In the `README.md` I find

> If you would like to contribute to elfe3D, report issues with elfe3D or seek support, please send an email to [p.rulff@tudelft.nl](mailto:p.rulff@tudelft.nl).

Whilst this provides a way for users to find out how to contribute, report issues, or seek support, I feel that having to get in touch by email doesn't exactly encourage open source contributions. I understand that adding new features and refactoring the code are things that you would want to discuss with potential contributors before they get started, to make sure the changes are consistent with your plans. However, I feel that the issue reporting process shouldn't require emailing (just open an issue on GitHub), nor should contributing small code changes and bug fixes. Would you be willing to change the process slightly and to mention this briefly in the contributing section?
