### 2025-04-16 emsig/elfe3D#3 @emgeopaula
Dear @jwallwork23,

thanks for pointing this out. I do, of course, want to encourage to use the issues reporting option of GitHub. I followed your suggestion and changed the corresponding lines in the readme.
