### 2025-04-16 emsig/elfe3D#3 @joewallwork
Thanks! Please could you also provide some brief guidelines on how to contribute? i.e., Open a pull request. And mention anything that contributors should do when preparing their contribution, e.g., check compiles, run tests, update documentation.
