### 2025-04-17 emsig/elfe3D#4 @joewallwork
[JOSS Review]: Dependencies list
Towards https://github.com/openjournals/joss-reviews/issues/7949.

From the JOSS review checklist:

> Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

While there are installation instructions provided in text form in `elfe3D/README.md`, please could you provide a clear list of the dependencies at the start of the installation instructions, before going into the detail of how to install them.
