### 2025-04-17 emsig/elfe3D#5 @joewallwork
[JOSS Review] (optional): Automated package management
Related to https://github.com/openjournals/joss-reviews/issues/7949.

From the JOSS review checklist:

> Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Whilst I'm also a Ubuntu user, I'm concerned that the installation instructions are quite limited to Ubuntu. I've been trying to build `elfe3D` on my laptop (Dell Latitude running Ubuntu 24.04) *without* system-wide package installs. I'd prefer not to (for example) `apt install tetgen` and instead use an automated package management solution. Note the following from the JOSS review checklist:

I've been trying to set up Spack to work with the repo and am using the configuration file
```yaml
spack:
  packages:
    all:
      compiler: ['gcc@11:']
  specs:
  - gmake
  - metis
  - mumps+metis~mpi+scotch
  - scotch+esmumps
  - tetgen
  concretizer:
    unify: true
```
Whilst this installs the dependencies correctly, I've not yet been able to hook this up in the `Makefile` yet.

I don't expect you to overhaul the build system for the purposes of the JOSS review, but I do feel that it would be beneficial to provide an installation approach using this or another automated package solution going forward.
