### 2025-04-17 emsig/elfe3D#6 @joewallwork
[JOSS Review]: Docstrings
Towards https://github.com/openjournals/joss-reviews/issues/7949.

It's great that `elfe3D` includes a manual and PhD thesis, as well as API documentation that can be generated using Doxygen. However, I feel that the API documentation isn't as useful as it could be because the docstrings don't give any information about the input variables. For example, `calculate_elemental_residuals` has input arguments `m`, `w`, `s`, etc. and it's unclear what these variables represent and what values should be passed.

It'd be great if variable descriptions could be added to the docstrings to improve the API documentation.
