### 2025-04-17 emsig/elfe3D#7 @joewallwork
[JOSS Review] (optional): Testing
Towards https://github.com/openjournals/joss-reviews/issues/7949.

From JOSS Review checklist:
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Whilst `elfe3D` includes an example, it would be beneficial to develop a test suite that can test the functionality more broadly. This could take the form of more examples/integration tests, a suite of unit tests, or both, for example.

Given that there is an example that demonstrates how to configure and run the main `elfe3D` program, I think this is sufficient for the JOSS Review criteria. However, it would benefit the package if the examples/tests were developed further.
