### 2025-04-24 emsig/elfe3D#4 @joewallwork
It'd also be helpful to note that MUMPS requires `sudo apt install libmetis-dev libparmetis-dev libscotch-dev libptscotch-dev libatlas-base-dev openmpi-bin libopenmpi-dev liblapack-dev libscalapack-openmpi-dev`. This is included as a comment in the MUMPS `Makefile`, but it'd be good to mention in the elfe3D installation instructions as well for clarity.
