### 2025-04-24 26ca35b
Update README.md
Guidelines for contributing
