### 2025-04-24 emsig/elfe3D#3 @emgeopaula
I detailed the information in the main README.md on how to contribute to elfe3D.
