+++ elfe3D/README.md
@@ -26,6 +26,18 @@ Please have a look at pages 34-40 in [Rulff (2023)](#references).
 
 ## Getting Started
 
+### Dependencies
+
+To install `elfe3D` on a Linux system, you need to have the following installed:
+
+- a modern Fortran compiler (2008 standard)
+- OpenBLAS
+- make
+- tetgen
+- MUMPS
+
+As the mesh generator `tetgen` and the solver `MUMPS` are open source packages. Installation instructions are given below. Note that during the installation of `MUMPS`, additional libaries have to be installed. Installation instructions are given below and in the `MUMPS` Makefile.
+
 ### Installation
 
 `elfe3D` can be compiled with `gfortran` or `ifort`. The provided Makefile is based on `gfortran` compilation. `gfortran` should be part of your system in Gnu compiler collection (gcc). Also, `OpenBLAS` and `make` packages are required. The latest compilation was performed on Ubuntu 22.04.
