+++ elfe3D/README.md
@@ -33,10 +33,10 @@ To install `elfe3D` on a Linux system, you need to have the following installed:
 - a modern Fortran compiler (2008 standard)
 - OpenBLAS
 - make
-- tetgen
+- TetGen
 - MUMPS
 
-As the mesh generator `tetgen` and the solver `MUMPS` are open source packages. Installation instructions are given below. Note that during the installation of `MUMPS`, additional libaries have to be installed. Installation instructions are given below and in the `MUMPS` Makefile.
+As the mesh generator `TetGen` and the solver `MUMPS` are open source packages. Installation instructions are given below. Note that during the installation of `MUMPS`, additional libaries have to be installed. Installation instructions are given below and in the `MUMPS` Makefile.
 
 ### Installation
 
@@ -44,7 +44,7 @@ As the mesh generator `tetgen` and the solver `MUMPS` are open source packages. 
 
 The following steps guide you through the `elfe3D` compilation:
 
-- `tetgen`: The open source mesh generator `tetgen` must be installed. It can be downloaded from <https://wias-berlin.de/software/index.jsp?id=TetGen> or directly installed via typing in your terminal:
+- `TetGen`: The open source mesh generator `TetGen` must be installed. It can be downloaded from <https://wias-berlin.de/software/index.jsp?id=TetGen> or directly installed via typing in your terminal:
 
   ``` bash
   $ sudo apt install tetgen
