### 2025-04-24 emsig/elfe3D#4 @emgeopaula
I added both suggestions to the manual.
