### 2025-04-24 emsig/elfe3D#7 @emgeopaula
Thanks for your suggestion! I am working on the development of further examples and tests.
