### 2025-04-24 emsig/elfe3D#5 @emgeopaula
I agree. I am currently working on setting up an automated package management solution to install `elfe3D`. Will upload our solution in the near future.
