### 2025-04-24 emsig/elfe3D#4 @joewallwork
Thanks! These additions look good to me so I'm happy to close this issue.
