### 2025-04-24 emsig/elfe3D#6 @emgeopaula
Thanks for suggesting this. The Doxygen documentation does indeed not directly give detailed information on the arguments passed to subroutines, but it does provide information on the variable types. Just below the input variable list, I can click on a link that brings me directly to the variable definitions in the corresponding source code lines. I agree that the variable descriptions can be commented in more detail in some subroutines and I will work on improving the code comments. However, all global variables are described in the Variable Definitions list at the beginning of elfe3D.f90, which is easily accessible via a similar link in the Doxygen File list. I would like to omit adding more detail to the Doxygen, as I feel that possible contributors will rather have a detailed look into the code itself and the thesis. Users can simply look into the manual, where they find necessary information about input variables.
Let me know what you think!
