+++ docs/README.md
@@ -3,3 +3,5 @@ Here, you find the elfe3D code documentation as pdf and the Kappa (thesis withou
 To generate a more detailed code documentation from source code comments, including information about routines, functions, and variables, you can run doxygen with the provided Doxyfile. For more information see: https://www.doxygen.nl/.
 
 For convenience, you find html files of the Doxygen Documentation in elfe3d_v1.1.0_Doxygen_html.zip. When opening the index.html file, the documentation will appear in your browser window.
+
+All global variables of the code are defined and described in a `Variable Definitions` list at the beginning of the main source code file `elfe3D.f90`. You can access it by either opening `elfe3D.f90` itself or in the Doxygen documentation via the corresponding link to `elfe3D.f90` in the Doxygen `File List`. Variable descriptions in the subroutines are currently under improvement.
