### 2025-04-25 emsig/elfe3D#6 @emgeopaula
> > I agree that the variable descriptions can be commented in more detail in some subroutines and I will work on improving the code comments.
>
> That'd be great, thanks.
>
Will do.

> > However, all global variables are described in the Variable Definitions list at the beginning of elfe3D.f90, which is easily accessible via a similar link in the Doxygen File list.
>
> Oh, good point - I hadn't noticed that. Perhaps this could be mentioned in the `elfe3D/README.md` file?
>
I have included the information in `docs/README.md`.

> > I would like to omit adding more detail to the Doxygen, as I feel that possible contributors will rather have a detailed look into the code itself and the thesis. Users can simply look into the manual, where they find necessary information about input variables.
>
> Sure, that makes sense.
