### 2025-04-25 cdfa0a5
Update elfe3d.f90
