+++ elfe3D/mod_define_model.f90
@@ -53,9 +53,11 @@ contains
   !---------------------------------------------------------------------
   !> @brief'
   !> subroutine for defining solver
+  !>
   !> Define the following in input file:
+  !>
   !> solver
-  !> 1: PARDISO (currently not available!)
+  !> 1: PARDISO (currently not available!) 
   !> 2: MUMPS
   !---------------------------------------------------------------------
   subroutine define_solver (solver)
@@ -99,35 +101,47 @@ contains
   !---------------------------------------------------------------------
   !> @brief
   !> subroutine for defining refinement parameters
+  !>
   !> Define the following in input file:
   !> maxRefSteps, maxUnknowns, betaRef, accuracyTol, vtk , 
-  !> errorEst_method , refStrategy             0
+  !> errorEst_method , refStrategy
   !>
   !> Maximum number of refinement steps:
   !> maxRefSteps = 0 means no refinement
-  !> run several frequencies only without refinement
+  !>
   !> Maximum number of unknowns:
   !> maxUnknowns  
+  !>
   !> threshold for number of elements to be refined:
   !> betaRef
   !> (e.g. 0.9 = those with 90% highest error estimator)
+  !>
   !> desired accuracy tolerance < 1:
   !> accuracyTol
+  !>
   !> write .vtk files for paraview with error estimates (yes/no) = (1/0)
   !> vtk
+  !>
   !> Method for error estimation:
+  !>
   !> errorEst_method
+  !>
   !> residuals (1)
   !> residuals and face jumps J (2)
   !> residuals and face jumps J and H (3)
   !> face jumps J (4)
   !> face jumps H (5)
   !> face jumps J & H  (6)
+  !>
   !> Refinement strategy:
   !> refStrategy
+  !>
   !> constant quality factor (0)
+  !>
   !> maxRefSteps-1 on low quality mesh, last step high quality mesh (1)
+  !>
   !> increasing quality factor (2)
+  !>
   !> increasing quality factor on mesh with detailled subsurface anomaly
   !> (-T and -d  option added) (3)
   !---------------------------------------------------------------------
@@ -233,6 +247,7 @@ contains
   !---------------------------------------------------------------------
   !> @brief
   !> subroutine for defining mesh
+  !>
   !> Define the following in input file: model_file_name
   !---------------------------------------------------------------------
   subroutine define_mesh (NodeFile, EdgeFile, ElementFile, NeighFile, &
@@ -319,45 +334,61 @@ contains
   !> subroutine for defining size of the source
   !>
   !> Define the following in input file:
+  !>
   !> source_type
+  !>
   !> source start coordinates
+  !>
   !> source end coordinates
+  !>
   !> current_direction
+  !>
   !> source_moment
   !>
   !> Options for source_type:
+  !>
   !> HED_x (0), HED_y (1), loop_source (2), arbitrary HED_x (3), 
   !> arbitrary HED_y (4), straight_source_segment in any direction (5)
   !> segmented_line_source (6), segmented_loop_source (7)
+  !>
   !> PR comment:  only 6 or 7 would be sufficient
-
+  !>
   !> HED_x (0), HED_y (1) are straight HEDs in exactly either x- or 
   !> y- direction
+  !>
   !> loop_source (2) is a squared loop source defined by northern
   !> line-source and midpoint, segments exactly on x- and y- direction
   !> if you choose arbitrary HED, you have to use node-boundary 
   !> markers = 3 for all source nodes in your .poly-inputfile
   !> this will be the fastest option, but it does probably not work 
   !> for refinement!
+  !>
   !> straight_source_segment in any direction (5) every edge between 
   !> source start and source endpoint will be a source edge
+  !>
[…]
