### 2025-04-25 #7949 @jwallwork23
We discussed what would be required to address the API documentation requests on issue https://github.com/emsig/elfe3D/issues/6 and I'm satisfied that we came to a conclusion. I'll leave that issue open for now because @emgeopaula is going to continue improving the docstrings in general, but for the purposes of this review I'm now happy that the checklist has been completed. Good to go from me! Thanks for your quick responses @emgeopaula.
