### 2025-04-30 #7949 @emgeopaula
@AnjaliSandip
W.r.t.
'Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
Archive the release on Zenodo/figshare/etc and post the DOI here.'

I have not changed anything in the source code during the review process, but added & modified documentation files. Do I need to make a new software release and Zenodo archive?
