+++ paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-The controlled-source electromagnetic method is a geophysical technique that detects variations in electrical and magnetic material properties in the subsurface. The method is valuable for both environmental assessments and resource exploration. To evaluate electromagnetic data obtained over complex subsurface structures, precise three-dimensional numerical modeling software is required. This is particularly important for tasks related to survey design and for inversion routines that estimate subsurface models based on field data inputs. The forward modelling program `elfe3D` simulates synthetic data of frequency-domain controlled-source electromagnetic configurations for three-dimensional subsurface models. To optimise the balance between the size of the problems and the accuracy of the solutions, adaptive mesh refinement strategies are employed. The `elfe3D` program allows the user to define an arbitrary survey setup: The sensor locations, where the electric and magnetic field responses are calculated, can be placed at or below the subsurface, or in the air. This flexibility facilitates realistic modelling of surface-based, borehole and airborne controlled-source electromagnetic surveys. 
+The controlled-source electromagnetic method is a geophysical technique that detects variations in electric and magnetic material properties in the subsurface. The method is valuable for both environmental assessments and resource exploration. To evaluate electromagnetic data obtained over complex subsurface structures, precise three-dimensional numerical modeling software is required. This is particularly important for tasks related to survey design and for inversion routines that estimate subsurface models based on field data inputs. The forward modelling program `elfe3D` simulates synthetic data of frequency-domain controlled-source electromagnetic configurations for three-dimensional subsurface models. To optimise the balance between the size of the problems and the accuracy of the solutions, adaptive mesh refinement strategies are employed. The `elfe3D` program allows the user to define an arbitrary survey setup: The sensor locations, where the electric and magnetic field responses are calculated, can be placed at or below the subsurface, or in the air. This flexibility facilitates realistic modelling of surface-based, borehole and airborne controlled-source electromagnetic surveys. 
 
 
 # Statement of need
