### 2025-05-15 #7949 @emgeopaula
@AnjaliSandip
I made a new release on github that includes the latest changes from the review: https://github.com/emsig/elfe3D/releases/tag/v1.0.1
The associated Zenodo Archive: https://doi.org/10.5281/zenodo.13309721
