### 2025-05-27 #7949 @emgeopaula
> [@emgeopaula](https://github.com/emgeopaula) We recommend that the title and author list for the metadata in your Zenodo software archive exactly match those in your JOSS paper to create a cohesive archive, though this is not a strict requirement. Do you want to make this change?

Would it be fine if I update the paper title and subtitle to

elfe3D v1.0.1:
Modelling with the total electric field approach using finite elements in 3D

?
