### 2025-05-27 #7949 @emgeopaula
> [@emgeopaula](https://github.com/emgeopaula) Your paper has grown in length — could some of it be moved to your docs?

@kthyng
Thanks for pointing this out. I was quite strict with the word count in the initial version. However, @AnjaliSandip and @jwallwork23 requested some restructuring and additions to the paper. As a result, it has indeed grown in length.

Please let me know if you want me to shorten it again. If necessary, I can remove the 'Workflow overview' section. However, I think this section is helpful for readers who want to get a brief overview of the workflow steps.
