### 2025-05-30 #7949 @kthyng
> Would it be fine if I update the paper title and subtitle

Yes that sounds good

> However, I think this section is helpful for readers who want to get a brief overview of the workflow steps.

Ok that is fine.

However, please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "delaunay"
