+++ paper.bib
@@ -13,14 +13,14 @@
    doi = {10.1007/s10596-023-10247-w},
    journal = {Computational Geosciences},
    keywords = {2010,68-08,78-10,86-10,geo-electromagnetics,goal-oriented meshing,high-order discretizations,mathematics subject classification,metallic infrastructures,numerical solutions,parallel computations},
-   title = {Meshing strategies for 3d geo-electromagnetic modeling},
+   title = {Meshing strategies for 3d geo-electromagnetic modeling n the presence of metallic infrastructure},
    year = {2023},
 }
 
 
 @book{Jin2014,
     author = {Jin, Jianming},
-    title = {The Finite Element Method in Electromagnetics},
+    title = {The {F}inite {E}lement {M}ethod in {E}lectromagnetics},
     year = {2014},
     publisher = {Wiley-IEEE Press},
     edition = {3rd}
@@ -87,7 +87,7 @@
 
 
 @article{Ren2013,
-  title={A goal-oriented adaptive finite-element approach for plane wave 3-D electromagnetic modelling},
+  title={A goal-oriented adaptive finite-element approach for plane wave 3-{D} electromagnetic modelling},
   author={Ren, Zhengyong and Kalscheuer, Thomas and Greenhalgh, Stewart and Maurer, Hansruedi},
   journal={Geophysical Journal International},
   volume={194},
@@ -175,7 +175,7 @@ journal = {Geophysics},
 mendeley-groups = {Modelling,Octavio_Collaboration},
 number = {6},
 pages = {E277--E291},
-title = {{Large-scale 3D geoelectromagnetic modeling using parallel adaptive high-order finite element method}},
+title = {{Large-scale {3D} geoelectromagnetic modeling using parallel adaptive high-order finite element method}},
 volume = {80},
 year = {2015}
 }
@@ -207,7 +207,7 @@ year = {2023}
    journal = {Geophysics},
    keywords = {borehole geophysics,electromagnetics,finite element,inversion,modeling},
    pages = {E139-E152},
-   title = {3D borehole-to-surface and surface electromagnetic modeling and inversion in the presence of steel infrastructure},
+   title = {{3D} borehole-to-surface and surface electromagnetic modeling and inversion in the presence of steel infrastructure},
    volume = {85},
    url = {https://doi.org/10.1190/geo2019-0034.1},
    year = {2020},
@@ -231,7 +231,7 @@ year = {2023}
 
 @article{Si2015,
 author = {Si, Hang},
-title = {TetGen, a Delaunay-Based Quality Tetrahedral Mesh Generator},
+title = {TetGen, a {D}elaunay-{B}ased {Q}uality {T}etrahedral {M}esh {G}enerator},
 year = {2015},
 issue_date = {January 2015},
 publisher = {Association for Computing Machinery},
@@ -286,7 +286,7 @@ doi = {10.1080/08123985.2019.1580118},
 
 @article{Werthmuller2021,
     author = {Werthmüller, Dieter and Rochlitz, Raphael and Castillo-Reyes, Octavio and Heagy, Lindsey},
-    title = {Towards an open-source landscape for 3-D CSEM modelling},
+    title = {Towards an open-source landscape for {3-D} {CSEM} modelling},
     journal = {Geophysical Journal International},
     volume = {227},
     number = {1},
@@ -299,7 +299,7 @@ doi = {10.1080/08123985.2019.1580118},
 @article{Weiss2023,
 author = {Weiss, Michael and Neytcheva, Maya and Kalscheuer, Thomas},
 publisher = {Computational Geosciences},
-title = {{Iterative solution methods for 3D controlled-source electromagnetic forward modelling of geophysical exploration scenarios}},
[…]
