+++ README.md
@@ -1,9 +1,11 @@
-# elfe3D
+# elfe3D v1.0.1
 Modelling with the total **el**ectric field approach using **f**inite **e**lements in **3D**
 
 _About:_
 
 `elfe3D` is a 3D forward modelling code that can simulate electric and magnetic field responses from frequency-domain controlled-source electromagnetic geophysical setups. It uses tetrahedral meshes and first-order finite-element approximations. In addition, adaptive mesh refinement approaches are implemented.
+
+This release (v.1.0.1) contains updated documentation files and the associated JOSS paper.
 
 _Statement of need:_
 
