### 2025-06-02 #7949 @emgeopaula
> > Would it be fine if I update the paper title and subtitle
>
> Yes that sounds good
>

Done.

> However, please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "delaunay"

Thank you for this comment. I have improved the reference list.
