### 2025-06-03 #7949 @emgeopaula
> [@emgeopaula](https://github.com/emgeopaula) Sorry I still don't see the paper title? https://zenodo.org/records/15422415

@kthyng
The 'Modelling with the total electric field approach using finite elements in 3D' was in the subtitle of the Zenodo archive. I moved it to the main title now.
