### 2025-06-03 #7949 @emgeopaula
> [@emgeopaula](https://github.com/emgeopaula) Can you remove the formatting for your title?

@kthyng
I removed the formatting.
