### 2025-06-06 #7949 @emgeopaula
> Congratulations on your new publication [@emgeopaula](https://github.com/emgeopaula)! Many thanks to editor [@AnjaliSandip](https://github.com/AnjaliSandip) and to reviewers [@pankajkmishra](https://github.com/pankajkmishra) and [@jwallwork23](https://github.com/jwallwork23) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

Many thanks to all of you @kthyng @AnjaliSandip @pankajkmishra @jwallwork23 for your valuable suggestions, our discussions and this enjoyable review process!

> [@emgeopaula](https://github.com/emgeopaula) If you'd like to join JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/join!

I did :)
